fix(Api): do not depend on globals (#169)
The use of globals in the bridge Api broke a few use-cases:
    - loading the bridge Api from two different locations
    - requiring `cordova-electron` without setting the globals first

This change refactors the run module back to an "extension method" of
Api. That is, it has to be called with `this` set to an Api instance.
This pattern is also used for `build` and `run` and it is pretty common
among the other platforms as well.