Improved ipcMain.handle handler and ipcRenderer.invoke callbacks for cdv-plugin-exec
The goal was to allow passing to the error callback an unaltered copy of the argument, as with the previous implementation you have to throw an exception or return a rejected promise, and this is forwarded to the ipcRenderer.invoke as a string incapsulated in an Error object. See https://github.com/electron/electron/issues/24427

This patch adds a success and error callback parameter for the plugin action, that can be used to implement the action function in a way more coherent with the Cordova architecture, and allowing to bypass the error callback parameter issue.

However, for backward compatibility, the action must call the callback parameters before returning (or before fulfilling the returned promise), otherwise it will work as in previous 3.x cordova-electron implementations, calling the success callback with the returned value or calling the error callback with an Error object.

Finally, _cdvElectronIPC.exec checks if success and error parameters are actually functions before calling them, in order to avoid annoying error log message if they were not defined.