Improved ipcMain.handle handler and ipcRenderer.invoke callbacks for cdv-plugin-exec
The main goal was to allow passing to the error callback an unaltered copy of the argument, as with the previous implementation you have to throw an exception or return a rejected promise, and this is forwarded to the ipcRenderer.invoke as a string encapsulated in an Error object. See https://github.com/electron/electron/issues/24427

Moreover, with the previous implementation, args for the action were encapsulated twice in an Array ( see https://github.com/apache/cordova-electron/issues/214 ). To make it more coherent with Cordova standards, and to allow backward support easily with a simple check, the first and the second arguments are success and error functions, and the third argument is the args array. So if the first argument is a function, we are using this cordova-electron~4 proposal, otherwise we are using cordova-electron~3, and it is the "arguments" array.

Finally, _cdvElectronIPC.exec checks if success and error parameters are actually functions before calling them, in order to avoid annoying error log messages if they were not defined.